appium-tests/pyproject.toml
[project]
name = "appium-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "appium-python-client>=5.0.0",
    "asgiref>=3.8.1",
    "pytest>=8.3.5",
    "qemu-qmp>=0.0.3",
]

[tool.pytest.ini_options]
markers = [
    "qemu: Run tests that require QEMU",
]

[tool.ruff]
line-length = 120
